new function
add function to check if we can convert text to float.  Then use that function to make sure the plus and minus function can insert minus sign only when the text is numeric or float.  This is not required right now because there isnt a situationt that would say disoplay text in current text box but in the future when i want error messages like cant devide by 0 to display i dont want the minus button to be able to add a minus to the text.