bug fixing
fixed some bugs having to do with setting focus when the app initialized, was causing abutton press to be interpreted as three button presses.  when i press 5 555 appears. this has to do with the fact that the focus is telling it to type as well as the root binds telling it to type right now the issue is half fixed with only two fives appearing instead of one. need to figure out a way to limit the user from setting focus.