added state changes
added state changes to every single function that affects current text and calculation text. This was the only way to change the text is for the button to me pressed any not through clicking on the text and free typing text. this prevents the bugs introduced by binding keys as well as limits the user from performing text injection. there is alot of added bloat but this is the only way i have found to get the behavior I want