lambda changes and inserted logic
made it lambda so that i could pass in arguments, then added the logic to the function to turn off or on the buttons of my choice. i still need to add status var into equal button so that it chan be changed and thus this function can check fo the change.