fixed issue
instead of keeping track of seprate string value we keep track of current_text value by creting a stringvar that is connect to current_text in innitiation. also added buttons to block.  fixed it so that pressing the remaining operating buttons will reset current_text and then add the number you pressed to it.