import change
change how arith is used. we are importing the module calc so that we can check calc.arith value and not the arith value that would of been created anew had we imported the variable. this was is refrences the original global variable in the calculation module. also made it so that other global variables are also checked this way