change arith format
instead of arith storing a bool, store a list of a bool.  This way when it is passed in as a variable it can be changed within the function without relying on global variables. this will make the code cleaner without the need to declare global variables in each function